# contracts/openapi/backend.openapi.json
{
  "components": {
    "schemas": {
...
      "AuthorRunArtifactSummary": {
        "properties": {
          "artifact_key": {
            "default": "",
            "title": "Artifact Key",
            "type": "string"
          },
          "artifact_type": {
            "title": "Artifact Type",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "artifact_type",
          "updated_at"
        ],
        "title": "AuthorRunArtifactSummary",
        "type": "object"
      },
      "AuthorRunCreateRequest": {
        "properties": {
          "raw_brief": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Raw Brief",
            "type": "string"
          }
        },
        "required": [
          "raw_brief"
        ],
        "title": "AuthorRunCreateRequest",
        "type": "object"
      },
      "AuthorRunCreateResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "review_ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "run_id",
          "status",
          "created_at"
        ],
        "title": "AuthorRunCreateResponse",
        "type": "object"
      },
      "AuthorRunEventPayload": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "node_name": {
            "title": "Node Name",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "event_id",
          "node_name",
          "event_type",
          "created_at"
        ],
        "title": "AuthorRunEventPayload",
        "type": "object"
      },
      "AuthorRunEventsResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuthorRunEventPayload"
            },
            "title": "Events",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "AuthorRunEventsResponse",
        "type": "object"
      },
      "AuthorRunGetResponse": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/AuthorRunArtifactSummary"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Node"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "raw_brief": {
            "title": "Raw Brief",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "review_ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "story_id",
          "status",
          "raw_brief",
          "created_at",
          "updated_at"
        ],
        "title": "AuthorRunGetResponse",
        "type": "object"
      },
      "AuthorStoryGetResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "draft_pack": {
            "additionalProperties": true,
            "title": "Draft Pack",
            "type": "object"
          },
          "latest_published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published At"
          },
          "latest_published_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Version"
          },
          "latest_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorRunGetResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "title",
          "created_at"
        ],
        "title": "AuthorStoryGetResponse",
        "type": "object"
      },
      "AuthorStoryListItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "latest_published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published At"
          },
          "latest_published_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Version"
          },
          "latest_run_current_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Current Node"
          },
          "latest_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Id"
          },
          "latest_run_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Status"
          },
          "latest_run_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Updated At"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "title",
          "created_at"
        ],
        "title": "AuthorStoryListItem",
        "type": "object"
      },
      "AuthorStoryListResponse": {
        "properties": {
          "stories": {
            "items": {
              "$ref": "#/components/schemas/AuthorStoryListItem"
            },
            "title": "Stories",
            "type": "array"
          }
        },
        "title": "AuthorStoryListResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StoryCreateRequest": {
        "properties": {
          "pack_json": {
            "additionalProperties": true,
            "title": "Pack Json",
            "type": "object"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "StoryCreateRequest",
        "type": "object"
      },
      "StoryCreateResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "status": {
            "const": "draft",
            "title": "Status",
            "type": "string"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "status",
          "created_at"
        ],
        "title": "StoryCreateResponse",
        "type": "object"
      },
...
      "StoryGetResponse": {
        "properties": {
          "pack": {
            "additionalProperties": true,
            "title": "Pack",
            "type": "object"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "story_id",
          "version",
          "pack"
        ],
        "title": "StoryGetResponse",
        "type": "object"
      },
      "StoryListItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "has_draft": {
            "title": "Has Draft",
            "type": "boolean"
          },
          "latest_published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published At"
          },
          "latest_published_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Version"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "title",
          "created_at",
          "has_draft"
        ],
        "title": "StoryListItem",
        "type": "object"
      },
      "StoryListResponse": {
        "properties": {
          "stories": {
            "items": {
              "$ref": "#/components/schemas/StoryListItem"
            },
            "title": "Stories",
            "type": "array"
          }
        },
        "title": "StoryListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "RPG Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/author/runs": {
      "post": {
        "operationId": "create_author_run_endpoint_author_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorRunCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Author Run Endpoint",
        "tags": [
          "author"
        ]
      }
    },
    "/author/runs/{run_id}": {
      "get": {
        "operationId": "get_author_run_endpoint_author_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorRunGetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Author Run Endpoint",
        "tags": [
          "author"
        ]
      }
    },
    "/author/runs/{run_id}/events": {
      "get": {
        "operationId": "get_author_run_events_endpoint_author_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorRunEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Author Run Events Endpoint",
        "tags": [
          "author"
        ]
      }
    },
    "/author/stories": {
      "get": {
        "operationId": "list_author_stories_endpoint_author_stories_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorStoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Author Stories Endpoint",
        "tags": [
          "author"
        ]
      }
    },
    "/author/stories/{story_id}": {
      "get": {
        "operationId": "get_author_story_endpoint_author_stories__story_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorStoryGetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Author Story Endpoint",
        "tags": [
          "author"
        ]
      }
    },
    "/author/stories/{story_id}/runs": {
      "post": {
        "operationId": "rerun_author_story_endpoint_author_stories__story_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorRunCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rerun Author Story Endpoint",
        "tags": [
          "author"
        ]
      }
    },
...
    "/stories": {
      "get": {
        "operationId": "list_stories_endpoint_stories_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Stories Endpoint",
        "tags": [
          "stories"
        ]
      },
      "post": {
        "operationId": "create_story_endpoint_stories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Story Endpoint",
        "tags": [
          "stories"
        ]
      }
    },
    "/stories/{story_id}": {
      "get": {
        "operationId": "get_story_endpoint_stories__story_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryGetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Story Endpoint",
        "tags": [
          "stories"
        ]
      }
    },
...
  }
}